refactor(rest): extract admin plugin-job schedule routes into routes/admin_plugin_jobs.py
Moves the five admin plugin-job schedule routes (GET/POST
plugin-jobs/schedules, PATCH/DELETE .../{schedule_id}, POST .../run) out
of create_app's closure into routes/admin_plugin_jobs.py.

plugin_schedule_fire (and SCHEDULE_FIRE_TOKEN/plugin_schedule_options)
move with them as a RestContext-taking function, since the plugin-job
scheduler tick (_plugin_schedule_loop, still in the closure) also calls
it. Collapses the closure's _SystemUser/_validate_cron/_next_fire
duplicate definitions into aliases onto routes/deps.py's SystemUser/
validate_cron/next_fire (added for the audit-schedules extraction) now
that both this group and that one need the same three.
